Files parser rewrite WIP
